Remove extraneous code
The contents of these files are either in need of a heavy (maybe total) rewrite, or deemed to be outside the scope of what Aurora is intended to be. Two large criteria which made that determination was if the code used Three.js and if the code was too opinionated-and-boilerplate-y.